fmt + clippy cleanups (CI -Dwarnings)
cargo fmt on the existing scaffold (CI's fmt --check rejected pre-fmt
code). Applied cargo clippy --fix for io::Error::other, useless vec!,
unneeded Copy-type .clone(), useless u16-as-u16 cast, and friends.
Plus the two non-auto-fixable ones:

- catalog.rs test helper: dropped redundant parens around assignment.
- ForkReader: added is_empty() to satisfy clippy::len_without_is_empty.

No behavior change.

+++ crates/fs-core/src/hfsplus/catalog.rs
@@ -266,7 +266,7 @@ mod tests {
 
     fn key_bytes(parent_id: u32, name: &str) -> Vec<u8> {
         let units: Vec<u16> = name.encode_utf16().collect();
-        let key_len: u16 = (4 + 2 + units.len() as u16 * 2) as u16;
+        let key_len: u16 = 4 + 2 + units.len() as u16 * 2;
         let mut out = Vec::new();
         out.extend_from_slice(&key_len.to_be_bytes());
         out.extend_from_slice(&parent_id.to_be_bytes());
